Avoid depending on checkout in masterbar
This fixes a performance problem introduced in #44206.
That PR added a number of dependencies to the critical path, by adding
them as dependencies to the masterbar.

This PR fixes things by isolating and loading the checkout parts
asynchronously, only when needed.
